backend/comp3900_stockoverflow/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-rest-knox==4.1.0

##npm install bootstrap jquery axios

